refactor Spy into FunctionSpy
