+++ core/_spying.ts
@@ -2,14 +2,12 @@ import { PropertySpy } from "./spying/property-spy";
 import { SpyCall } from "./spying/spy-call";
 import { SpyOn } from "./spying/spy-on";
 import { SpyOnProperty } from "./spying/spy-on-property";
-import { Spy } from "./spying/spy";
 import { FunctionSpy } from "./spying/function-spy";
 
 export {
-    PropertySpy,
-    SpyCall,
-    SpyOn,
-    SpyOnProperty,
-    Spy,
-    FunctionSpy
+   PropertySpy,
+   SpyCall,
+   SpyOn,
+   SpyOnProperty,
+   FunctionSpy
 };
