--- test/unit-tests/spying/spy/and-return.spec.ts
@@ -1,31 +1,0 @@
-import { Spy } from "../../../../core/spying/spy";
-import { Expect, Test, TestCase, SpyOn } from "../../../../core/alsatian-core";
-
-export class ReturnTests {
-
-  @TestCase(undefined)
-  @TestCase(null)
-  @TestCase(0)
-  @TestCase(42)
-  @TestCase(4.2)
-  @TestCase(-4.2)
-  @TestCase("")
-  @TestCase("something")
-  public givenReturnValueIsReturned(expectedReturnValue: any) {
-    let object = {
-      originalFunction: () => {}
-    };
-
-    SpyOn(object, "originalFunction");
-
-    let originalFunction = object.originalFunction;
-
-    let spy = new Spy(object.originalFunction, object);
-
-    spy.andReturn(expectedReturnValue);
-
-    let actualReturnValue = spy.call([]);
-
-    Expect(actualReturnValue).toBe(expectedReturnValue);
-  }
-}
