+++ tsconfig.json
@@ -86,7 +86,6 @@
         "./core/spying/spy-call.ts",
         "./core/spying/spy-on-property.ts",
         "./core/spying/spy-on.ts",
-        "./core/spying/spy.ts",
         "./core/test-fixture.ts",
         "./core/test-loader.ts",
         "./core/test-output.ts",
@@ -139,17 +138,17 @@
         "./test/unit-tests/results/test-fixture-results.spec.ts",
         "./test/unit-tests/results/test-results.spec.ts",
         "./test/unit-tests/results/test-set-results.spec.ts",
+        "./test/unit-tests/spying/function-spy/and-call-through.spec.ts",
+        "./test/unit-tests/spying/function-spy/and-call.spec.ts",
+        "./test/unit-tests/spying/function-spy/and-return.spec.ts",
+        "./test/unit-tests/spying/function-spy/and-stub.spec.ts",
+        "./test/unit-tests/spying/function-spy/call.spec.ts",
         "./test/unit-tests/spying/function-spy/restore.spec.ts",
         "./test/unit-tests/spying/property-spy/and-return-value.spec.ts",
         "./test/unit-tests/spying/property-spy/constructor.spec.ts",
         "./test/unit-tests/spying/spy-call.spec.ts",
         "./test/unit-tests/spying/spy-on-property.spec.ts",
         "./test/unit-tests/spying/spy-on.spec.ts",
-        "./test/unit-tests/spying/spy/and-call-through.spec.ts",
-        "./test/unit-tests/spying/spy/and-call.spec.ts",
-        "./test/unit-tests/spying/spy/and-return.spec.ts",
-        "./test/unit-tests/spying/spy/and-stub.spec.ts",
-        "./test/unit-tests/spying/spy/call.spec.ts",
         "./test/unit-tests/test-fixture/get-tests.spec.ts",
         "./test/unit-tests/test-loader/load-test-fixture.spec.ts",
         "./test/unit-tests/test-loader/load-test-fixtures-tests/default-export-fixture.spec.ts",
@@ -167,7 +166,8 @@
         "./test/unit-tests/test-runner/no-tests-error.spec.ts",
         "./test/unit-tests/test-runner/pre-test.spec.ts",
         "./test/unit-tests/test-set/load-test.spec.ts",
-        "./scripts/tag-release.ts"
+        "./scripts/tag-release.ts",
+        "./typings/index.d.ts"
     ],
     "atom": {
         "rewriteTsconfig": true
